pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "finance"
version = "0.0.1"
description = "Implement finances analysis"
authors = [
  { name = "Paulo Beckhauser", email = "your@email.com" },
]
license = { file = "LICENSE" }
readme = "README.md"
requires-python = ">=3.11"
dynamic = ["optional-dependencies"]
dependencies = [
    "cookiecutter>=2.6.0",
    "hydra-core>=1.3.2",
    "investpy>=1.0.8",
    "lightning[extra]>=2.5.0.post0",
    "matplotlib>=3.10.0",
    "pip-chill>=1.0.3",
    "pipreqs>=0.5.0",
    "pytorch-forecasting>=1.2.0",
    "pytorch-lightning>=2.5.0.post0",
    "scikit-learn>=1.6.0",
    "torch>=2.5.1",
    "tqdm>=4.67.1",
    "transformers>=4.47.1",
    "typer>=0.15.1",
    "yfinance>=0.2.51",
]

[tool.setuptools.dynamic]
dependencies = {file = ["requirements.txt"]}

[tool.setuptools.dynamic.optional-dependencies]
dev = {file = ['requirements_dev.txt']}

[tool.ruff]
line-length = 120

[tool.coverage.run]
omit = ["tests/*"]
